examples/binarization/requirements.txt
gdown
Pillow
numpy
iunets
